+++ src/main/java/com/repository/IListingRepository.java
@@ -1,10 +1,10 @@
 package src.main.java.com.repository;
 
-import src.main.java.com.model.Listing;
 import java.util.List;
+import src.main.java.com.model.Listing;
 
 public interface IListingRepository {
-    int add(String title, String description, double price, String username, String category);
+    int add(String title, String description, int price, String username, String category);
     Listing get(int listingId);
     boolean delete(int listingId);
     List<Listing> getByCategory(String category);
